+++ packages/mobile/package.json
@@ -44,7 +44,7 @@
     "jsdom": "^22.1.0",
     "terser": "^5.4.0",
     "typescript": "^5.1.6",
-    "vite": "^5.0.0",
+    "vite": "5.0.4",
     "vitest": "^0.34.6"
   },
   "description": "An Ionic project"
