+++ packages/mobile/src/components/ProjectListItem.tsx
@@ -1,7 +1,12 @@
 import {
   IonItem,
   IonLabel,
-  IonNote
+  IonNote,
+  IonCard,
+  IonCardHeader,
+  IonCardTitle,
+  IonCardSubtitle,
+  IonCardContent
   } from '@ionic/react';
 import { Project } from '../data/projects';
 import './ProjectListItem.css';
@@ -13,19 +18,15 @@ interface ProjectListItemProps {
 
 const ProjectListItem: React.FC<ProjectListItemProps> = ({ project }) => {
   return (
-    <IonItem routerLink={`/project/${project.id}`} detail={false}>
-      <div slot="start" className="dot"></div>
-      <IonLabel className="ion-text-wrap">
-        <h2>
-          {project.title}
-          <span className="date">
-            <IonNote style={{ color: getStatusColor(project.status) }}>{getStatus(project.status)}</IonNote>
-          </span>
-        </h2>
-        {/* <h3>{project.subject}</h3> */}
-        <p>{project.description}</p>
-      </IonLabel>
-    </IonItem>
+    <IonCard routerLink={`/project/${project.id}`} style={{ padding: '10px' }}>
+      <img alt="Silhouette of mountains" src={`https://source.unsplash.com/random/10×5/?${project.Country}`} style={{ padding: '20px' }}/>
+      <IonCardHeader>
+        <IonCardTitle>{project.ProjectName}</IonCardTitle>
+        <IonCardSubtitle style={{ color: getStatusColor(project.status) }}>{getStatus(project.status)}</IonCardSubtitle>
+      </IonCardHeader>
+
+      <IonCardContent>{project.FullDescription}</IonCardContent>
+    </IonCard>
   );
 };
