+++ packages/mobile/src/data/projects_0.ts
@@ -1,0 +1,64 @@
+export interface Project {
+  fromName: string;
+  title: string;
+  description: string,
+  subject: string;
+  date: string;
+  status: number;
+  id: number;
+}
+
+//'0.Por Aprobar', '1.Aprobado', '2.En ejecución','-1.Cancelado'
+
+const projects: Project[] = [
+  {
+    fromName: 'Matt Chorsey',
+    title: 'Project 1',
+    description: `1 Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.`,
+    subject: 'New event: Trip to Vegas',
+    date: '9:32 AM',
+    status: 2,
+    id: 0
+  },
+  {
+    fromName: 'Lauren Ruthford',
+    title: 'Project 2',
+    description: `2 Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.`,
+    subject: 'Long time no chat',
+    date: '6:12 AM',
+    status: 0,
+    id: 1
+  },
+  {
+    fromName: 'Jordan Firth',
+    title: 'Project 3',
+    description: `3 Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.`,
+    subject: 'Report Results',
+    date: '4:55 AM',
+    status: 2,
+    id: 2
+
+  },
+  {
+    fromName: 'Bill Thomas',
+    title: 'Project 4',
+    description: `4 Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.`,
+    subject: 'The situation',
+    date: '9:00 AM',
+    status: 1,
+    id: 3
+  },
+  {
+    fromName: 'John Doe',
+    title: 'Project 5',
+    description: `5 Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.`,
+    subject: 'The statement',
+    date: 'Yesterday',
+    status: -1,
+    id: 4
+  }
+];
+
+export const getProjects = () => projects;
+
+export const getProject = (id: number) => projects.find(p => p.id === id);
